+++ docs/history.rst
@@ -8,6 +8,7 @@ Version History
 - Add :func:`sprockets.mixins.mediatype.content.add_transcoder`.
 - Add :class:`sprockets.mixins.mediatype.transcoders.JSONTranscoder`
 - Add :class:`sprockets.mixins.mediatype.transcoders.MsgPackTranscoder`
+- Add :class:`sprockets.mixins.mediatype.transcoders.BinaryWrapper`
 
 `1.0.4`_ (14 Sep 2015)
 ----------------------